Add support for "technical" wagons
That is, ones that are only wagons internally, for modding API limitation reasons. In particular, these are almost certainly omnidirectional, and this handles that.

As before, I have tested it with both my technical wagon/chest/thing and ensured that it did not break vanilla ones.

The current "technical" check is a friction of 999999 - absurdly large, and fairly specific - but if [my request for the "max_speed" variable access](https://forums.factorio.com/viewtopic.php?f=28&t=47087&p=333859#p333859) gets implemented, checking a max_speed of 0 on the LuaEntityPrototype is a cleaner choice.